requirements.txt
numpy
pandas
Pillow
matplotlib
jinja2
torch
transformers
safetensors
sentence-transformers
datasets
evaluate
rich
